test(textarea): migrate highlight
