+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-no-fill-invalid-md-ltr-Mobile-Firefox-linux.png
