+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-no-fill-valid-ios-ltr-Mobile-Firefox-linux.png
