+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-outline-focus-ios-ltr-Mobile-Chrome-linux.png
