+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-outline-invalid-ios-ltr-Mobile-Safari-linux.png
