+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-outline-invalid-md-ltr-Mobile-Chrome-linux.png
