+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-outline-valid-ios-ltr-Mobile-Chrome-linux.png
