+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-outline-valid-md-ltr-Mobile-Safari-linux.png
