+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-solid-focus-md-ltr-Mobile-Safari-linux.png
