+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-solid-invalid-ios-ltr-Mobile-Chrome-linux.png
