+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-solid-invalid-md-ltr-Mobile-Chrome-linux.png
