+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-solid-valid-ios-ltr-Mobile-Chrome-linux.png
