+++ core/src/components/textarea/test/highlight/textarea.e2e.ts-snapshots/textarea-solid-valid-md-ltr-Mobile-Safari-linux.png
